# swagger.json
{
	"components": {
...
		"schemas": {
			"IRefItem": {
				"properties": {
					"name": {
						"type": "string"
					},
					"version": {
						"type": "number",
						"format": "double"
					},
					"lastUpdate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"name",
					"version",
					"lastUpdate"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ISelectorsMeta": {
				"properties": {
					"ref": {
						"$ref": "#/components/schemas/IRefItem"
					}
				},
				"required": [
					"ref"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SelectorTypes": {
				"enum": [
					"menu-category",
					"schema-category"
				],
				"type": "string"
			},
			"ISelectorContentsItem": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"resources": {
						"properties": {
							"icon": {
								"type": "string",
								"nullable": true
							},
							"main": {
								"type": "string",
								"nullable": true
							}
						},
						"required": [
							"icon",
							"main"
						],
						"type": "object"
					},
					"assets": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"extra": {
						"properties": {},
						"additionalProperties": {
							"additionalProperties": true
						},
						"type": "object",
						"nullable": true
					}
				},
				"required": [
					"name",
					"resources"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISelectorContents": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"anyOf": [
						{
							"$ref": "#/components/schemas/ISelectorContentsItem"
						},
						{
							"additionalProperties": true
						}
					]
				}
			},
			"ISelectorItem": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/SelectorTypes"
					},
					"active": {
						"type": "boolean"
					},
					"contents": {
						"$ref": "#/components/schemas/ISelectorContents"
					},
					"joint": {
						"type": "string"
					},
					"extra": {
						"properties": {},
						"additionalProperties": {
							"additionalProperties": true
						},
						"type": "object",
						"nullable": true
					}
				},
				"required": [
					"type",
					"active",
					"contents",
					"joint"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISelectorsResponse": {
				"properties": {
					"meta": {
						"$ref": "#/components/schemas/ISelectorsMeta"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/ISelectorItem"
						},
						"type": "array"
					},
					"error": {
						"items": {
							"properties": {
								"message": {
									"type": "string"
								},
								"code": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"message",
								"code"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ISelectorResponse": {
				"properties": {
					"meta": {
						"$ref": "#/components/schemas/ISelectorsMeta"
					},
					"data": {
						"$ref": "#/components/schemas/ISelectorItem"
					},
					"error": {
						"items": {
							"properties": {
								"message": {
									"type": "string"
								},
								"code": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"message",
								"code"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ISelectorCreateRequest": {
				"properties": {
					"active": {
						"type": "boolean"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/SelectorTypes"
					},
					"contents": {
						"$ref": "#/components/schemas/ISelectorContents"
					},
					"extra": {
						"properties": {},
						"additionalProperties": {
							"additionalProperties": true
						},
						"type": "object",
						"nullable": true
					}
				},
				"required": [
					"active",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISelectorUpdateRequest": {
				"properties": {
					"active": {
						"type": "boolean"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/SelectorTypes"
					},
					"contents": {
						"$ref": "#/components/schemas/ISelectorContents"
					},
					"extra": {
						"properties": {},
						"additionalProperties": {
							"additionalProperties": true
						},
						"type": "object",
						"nullable": true
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"clientAccessToken": {
				"type": "apiKey",
				"name": "x-authorization",
				"in": "header"
			},
			"accessToken": {
				"type": "apiKey",
				"name": "x-access-token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "tornado-server",
		"version": "1.1.0",
		"license": {
			"name": "Commercial"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/selectors": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISelectorsResponse"
								},
								"examples": {
									"example": {
										"value": {
											"meta": {
												"ref": {
													"name": "selectors",
													"version": 1,
													"lastUpdate": "2020-11-18T21:36:38.127Z"
												}
											},
											"data": [
												{
													"id": "507c7f79bcf86cd7994f6c0e",
													"active": true,
													"type": "menu-category",
													"contents": {
														"RU": {
															"name": "Selectors on concert",
															"description": "Lorem Ipsum is simply dummy text of the printing and typesetting industry.",
															"color": "#000000",
															"resources": {
																"main": "g8h07f79bcf86cd7994f9d7k",
																"icon": "k7h97f79bcf86cd7994f0i9e"
															},
															"assets": [
																"g8h07f79bcf86cd7994f9d7k"
															]
														}
													},
													"joint": "890c7f79bcf86cd7994f3t8y",
													"extra": {
														"key": "value"
													}
												}
											]
										}
									}
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Selector"
				],
				"security": [
					{
						"clientAccessToken": []
					},
					{
						"accessToken": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/SelectorTypes"
						}
					}
				]
			}
		},
		"/selector/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISelectorResponse"
								},
								"examples": {
									"example": {
										"value": {
											"meta": {
												"ref": {
													"name": "selectors",
													"version": 1,
													"lastUpdate": "2020-11-18T21:36:38.131Z"
												}
											},
											"data": {
												"id": "507c7f79bcf86cd7994f6c0e",
												"active": true,
												"type": "menu-category",
												"contents": {
													"RU": {
														"name": "Selectors on concert",
														"description": "Lorem Ipsum is simply dummy text of the printing and typesetting industry.",
														"color": "#000000",
														"resources": {
															"main": "g8h07f79bcf86cd7994f9d7k",
															"icon": "k7h97f79bcf86cd7994f0i9e"
														},
														"assets": [
															"g8h07f79bcf86cd7994f9d7k"
														]
													}
												},
												"joint": "890c7f79bcf86cd7994f3t8y",
												"extra": {
													"key": "value"
												}
											}
										}
									}
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Selector"
				],
				"security": [
					{
						"clientAccessToken": []
					},
					{
						"accessToken": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISelectorResponse"
								},
								"examples": {
									"example": {
										"value": {
											"meta": {
												"ref": {
													"name": "selectors",
													"version": 1,
													"lastUpdate": "2020-11-18T21:36:38.136Z"
												}
											},
											"data": {
												"id": "507c7f79bcf86cd7994f6c0e",
												"active": true,
												"type": "menu-category",
												"contents": {
													"RU": {
														"name": "Selectors on concert",
														"description": "Lorem Ipsum is simply dummy text of the printing and typesetting industry.",
														"color": "#000000",
														"resources": {
															"main": "g8h07f79bcf86cd7994f9d7k",
															"icon": "k7h97f79bcf86cd7994f0i9e"
														},
														"assets": [
															"g8h07f79bcf86cd7994f9d7k"
														]
													}
												},
												"joint": "890c7f79bcf86cd7994f3t8y",
												"extra": {
													"key": "value"
												}
											}
										}
									}
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Selector"
				],
				"security": [
					{
						"clientAccessToken": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISelectorUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISelectorResponse"
								},
								"examples": {
									"example": {
										"value": {
											"meta": {
												"ref": {
													"name": "selectors",
													"version": 1,
													"lastUpdate": "2020-11-18T21:36:38.139Z"
												}
											}
										}
									}
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Selector"
				],
				"security": [
					{
						"clientAccessToken": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/selector": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISelectorResponse"
								},
								"examples": {
									"example": {
										"value": {
											"meta": {
												"ref": {
													"name": "selectors",
													"version": 1,
													"lastUpdate": "2020-11-18T21:36:38.132Z"
												}
											},
											"data": {
												"id": "507c7f79bcf86cd7994f6c0e",
												"active": true,
												"type": "menu-category",
												"contents": {
													"RU": {
														"name": "Selectors on concert",
														"description": "Lorem Ipsum is simply dummy text of the printing and typesetting industry.",
														"color": "#000000",
														"resources": {
															"main": "g8h07f79bcf86cd7994f9d7k",
															"icon": "k7h97f79bcf86cd7994f0i9e"
														},
														"assets": [
															"g8h07f79bcf86cd7994f9d7k"
														]
													}
												},
												"joint": "890c7f79bcf86cd7994f3t8y",
												"extra": {
													"key": "value"
												}
											}
										}
									}
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Selector"
				],
				"security": [
					{
						"clientAccessToken": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISelectorCreateRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}